requirements.txt
asgiref>=3.6.0
black>=23.1.0
certifi>=2022.12.7
charset-normalizer>=3.0.1
click>=8.1.3
coreapi>=2.3.3
coreschema>=0.0.4
Django>=4.1.6
django-stubs>=1.14.0
django-stubs-ext>=0.7.0
djangorestframework>=3.14.0
djangorestframework-stubs>=1.8.0
drf-yasg>=1.21.4
drf-yasg-stubs>=0.1.4
idna>=3.4
inflection>=0.5.1
itypes>=1.2.0
Jinja2>=3.1.2
MarkupSafe>=2.1.2
mypy>=0.991
mypy-extensions>=1.0.0
packaging>=23.0
pathspec>=0.11.0
platformdirs>=2.6.2
psycopg2-binary>=2.9.5
python-dotenv>=0.21.1
pytz>=2022.7.1
requests>=2.28.2
ruamel.yaml>=0.17.21
sqlparse>=0.4.3
tomli>=2.0.1
types-pytz>=2022.7.1.0
types-PyYAML>=6.0.12.4
types-requests>=2.28.11.11
types-urllib3>=1.26.25.5
typing_extensions>=4.4.0
uritemplate>=4.1.1
urllib3>=1.26.14
